requirements.txt
httpx
httpx[brotli]
httpx[socks]
httpx-socks
python_telegram_bot
loguru
